pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dzira"
version = "0.2.1"
dependencies = [
    "click >= 8.1",
    "jira",
    "python-dotenv",
    "tabulate",
]
authors = [
    { name="Piotr Kaznowski", email="piotr@kazno.dev" },
]
description = "A cli wrapper for jira API to track sprints and manage worklogs"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.8",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development",
]

[project.scripts]
dzira = "dzira.cli.commands:main"

[project.urls]
Homepage = "https://github.com/caseneuve/dzira"
Issues = "https://github.com/caseneuve/dzira/issues"
Changelog = "https://github.com/caseneuve/dzira/blob/master/CHANGELOG.md"